requirements.txt
qrcode>=7.4.2
pillow>=10.4.0
